refactor: (#49) 디렉토리명 수정
