+++ src/main/java/org/dinosaur/foodbowl/domain/member/persistence/MemberRoleRepository.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.member.repository;
+package org.dinosaur.foodbowl.domain.member.persistence;
 
 import org.dinosaur.foodbowl.domain.member.domain.MemberRole;
 import org.springframework.data.repository.Repository;
