+++ src/main/java/org/dinosaur/foodbowl/domain/photo/persistence/ThumbnailRepository.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.photo.repository;
+package org.dinosaur.foodbowl.domain.photo.persistence;
 
 import org.dinosaur.foodbowl.domain.photo.domain.Thumbnail;
 import org.springframework.data.repository.Repository;
