+++ src/main/java/org/dinosaur/foodbowl/domain/store/persistence/CategoryRepository.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.store.repository;
+package org.dinosaur.foodbowl.domain.store.persistence;
 
 import org.dinosaur.foodbowl.domain.store.domain.Category;
 import org.springframework.data.repository.Repository;
