+++ src/main/java/org/dinosaur/foodbowl/domain/store/persistence/SchoolRepository.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.store.repository;
+package org.dinosaur.foodbowl.domain.store.persistence;
 
 import org.dinosaur.foodbowl.domain.store.domain.School;
 import org.springframework.data.repository.Repository;
