+++ src/main/java/org/dinosaur/foodbowl/domain/store/persistence/StoreSchoolRepository.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.store.repository;
+package org.dinosaur.foodbowl.domain.store.persistence;
 
 import org.dinosaur.foodbowl.domain.store.domain.StoreSchool;
 import org.springframework.data.repository.Repository;
